wip: refactor 2
