+++ work-creator/cmd/common.go
@@ -1,0 +1,12 @@
+package cmd
+
+import (
+	"k8s.io/client-go/kubernetes/scheme"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+func getClient() (client.Client, error) {
+	config := ctrl.GetConfigOrDie()
+	return client.New(config, client.Options{Scheme: scheme.Scheme})
+}
