+++ work-creator/cmd/update_status.go
@@ -1,0 +1,86 @@
+package cmd
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"path/filepath"
+
+	"github.com/spf13/cobra"
+	"github.com/syntasso/kratix/work-creator/lib"
+	"github.com/syntasso/kratix/work-creator/lib/helpers"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"sigs.k8s.io/yaml"
+)
+
+func updateStatusCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "update-status",
+		Short: "Update status of Kubernetes resources",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			ctx := context.Background()
+			return runUpdateStatus(ctx)
+		},
+	}
+
+	return cmd
+}
+
+func runUpdateStatus(ctx context.Context) error {
+	workspaceDir := "/work-creator-files"
+	statusFile := filepath.Join(workspaceDir, "metadata", "status.yaml")
+
+	params := helpers.GetParametersFromEnv()
+
+	client, err := helpers.GetK8sClient()
+	if err != nil {
+		return fmt.Errorf("failed to create Kubernetes client: %w", err)
+	}
+
+	objectClient := client.Resource(helpers.ObjectGVR(params)).Namespace(params.ObjectNamespace)
+
+	existingObj, err := objectClient.Get(ctx, params.ObjectName, metav1.GetOptions{})
+	if err != nil {
+		return fmt.Errorf("failed to get existing object: %w", err)
+	}
+
+	existingStatus := map[string]any{}
+	if existingObj.Object["status"] != nil {
+		existingStatus = existingObj.Object["status"].(map[string]any)
+	}
+
+	// Load incoming status.yaml if exists
+	incomingStatus, err := readStatusFile(statusFile)
+	if err != nil {
+		return fmt.Errorf("failed to load incoming status: %w", err)
+	}
+
+	mergedStatus := lib.MergeStatuses(existingStatus, incomingStatus)
+	if params.IsLastPipeline {
+		mergedStatus = lib.MarkAsCompleted(mergedStatus)
+	}
+
+	// Apply merged status to the existing object
+	existingObj.Object["status"] = mergedStatus
+
+	// Update the object's status
+	if _, err = objectClient.UpdateStatus(ctx, existingObj, metav1.UpdateOptions{}); err != nil {
+		return fmt.Errorf("failed to update status: %w", err)
+	}
+
+	return nil
+}
+
+func readStatusFile(statusFile string) (map[string]any, error) {
+	incomingStatus := map[string]any{}
+	if _, err := os.Stat(statusFile); err == nil {
+		incomingStatusBytes, err := os.ReadFile(statusFile)
+		if err != nil {
+			return nil, fmt.Errorf("failed to read status file: %w", err)
+		}
+		if err := yaml.Unmarshal(incomingStatusBytes, &incomingStatus); err != nil {
+			return nil, fmt.Errorf("failed to unmarshal incoming status: %w", err)
+		}
+	}
+	return incomingStatus, nil
+}
