+++ work-creator/lib/suite_test.go
@@ -1,4 +1,4 @@
-package pipeline_test
+package lib_test
 
 import (
 	"testing"
