+++ work-creator/lib/work_creator.go
@@ -1,4 +1,4 @@
-package pipeline
+package lib
 
 import (
 	"context"
