+++ work-creator/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/syntasso/kratix/work-creator/cmd"
+
+func main() {
+	cmd.Execute()
+}
