--- work-creator/pipeline/lib/lib_suite_test.go
@@ -1,13 +1,0 @@
-package lib_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestLib(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Lib Suite")
-}
