+++ work-creator/samples/complete-for-promise/input/baz/baz-namespace-resource-request.yaml
