+++ work-creator/samples/complete-for-promise/input/configmap.yaml
