+++ work-creator/samples/complete-for-promise/input/foo/bar/namespace-resource-request.yaml
