+++ work-creator/samples/complete-for-promise/input/foo/multi-resource-requests.yaml
