+++ work-creator/samples/complete-for-promise/metadata/destination-selectors.yaml
