+++ work-creator/samples/complete-updated/input/baz/baz-namespace-resource-request.yaml
