+++ work-creator/samples/complete-updated/input/configmap.yaml
