+++ work-creator/samples/complete-updated/input/foo/bar/namespace-resource-request.yaml
