+++ work-creator/samples/complete-updated/input/foo/multi-resource-requests.yaml
