+++ work-creator/samples/complete-updated/metadata/destination-selectors.yaml
