+++ work-creator/samples/complete/input/baz/baz-namespace-resource-request.yaml
