+++ work-creator/samples/complete/input/configmap.yaml
