+++ work-creator/samples/complete/input/foo/bar/namespace-resource-request.yaml
