+++ work-creator/samples/complete/input/foo/multi-resource-requests.yaml
