+++ work-creator/samples/complete/metadata/destination-selectors.yaml
