+++ work-creator/samples/empty-default-workload-group/input/baz/baz-namespace-resource-request.yaml
