+++ work-creator/samples/empty-default-workload-group/metadata/destination-selectors.yaml
