+++ work-creator/samples/empty-metadata/input/multi-resource-requests.yaml
