+++ work-creator/samples/empty-metadata/input/namespace-resource-request.yaml
